+++ features/scripts/build_unity_2018_abi_splits.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -e
+
+cd features/fixtures/unity_2018/example
+./gradlew clean uploadBugsnagUnityArmeabi-v7a-releaseMapping --stacktrace